New Zipfian distribution generator
This solution is much faster, but with this approach we could lose the
property of generating more edges to nodes in the top of the Zipfian
distribution when increasing the amount of nodes. This will be the case
when the all interface-connections of a node are full. However, this
solves the nmOfEdgesPerIteration problem for a Zipfian in- and
out-distribution