Update interface-connections Zipf each iteration
This fixes the problem where a node in the top of the Zipfian
distribution (one of the few nodes with a lot of connections) has a
bounded maximum number of connections.

This approach updates the amount of interface-connections for each node
in a Zipfian distribution in each iteration. The maximum number of
interface connections of a node in the top of the Zipfian distribution
is now bounded by the iterationNumber. This iterationNumber will grow
when the total number of nodes (i.e. the graph size) grows. This means
that the amount of connections to an node in the top of the Zipfian
distribution will increase when the graph grows.